A grammar for observation: watch/subscribe crossed with rows/changes
Two axes were being carried as flags on one method. Both are now in the
name, and the names are the ones this library already uses.

subscribe has meant "hold the feed open" here since long before rows had
names, and watch has meant "do not". That answers how a caller says to keep
the socket open: they say subscribe. from: "current" | "now" becomes the
noun, because a consumer that must process every row and one that only
reacts to change want opposite things and the wrong one is silent until a
backlog exists.

    one shot        query          queryRows
    local           watch          watchRows      watchChanges
    feed held       subscribe      subscribeRows  subscribeChanges

The left column is the projection tree a view wants; the right is the
identified rows a consumer wants. queryChanges is missing because a one-shot
delta is not a thing, which is a grammar that excludes the meaningless
rather than a list of methods.

Delivery is pull only. The library owns the listener, so it does nothing but
enqueue and the consumer's work happens on its own turn: neither working
inside the notification nor being abandoned at listenerTimeoutMs is
something a consumer has to remember. Push consumers already have watch and
subscribe. A stream is AsyncIterable, so `for await (const change of
stream)` is the whole interface, plus stop(), dropped and pending.

The queue keeps two buffers. What the read produced is never dropped,
because the read is the reliable path and a consumer that silently lost part
of its backlog could not notice. Notifications are bounded and drop oldest,
because back pressure there would block the save behind the listener.

Only options that tune stay options: capacity. Modes are names.

observeChanges and streamChanges are gone rather than deprecated; neither
has been released. queryRows keeps its name and its meaning.

31 tests. Three pin invariants by construction: listeners register before
the read (fails when reordered), a row saved during startup arrives exactly
once (fails without the suppression), and starting rows survive a capacity
of one (fails when routed through the bounded path). A recording Network
proves subscribeRows opens the feed and stop() releases it.
